requirements.txt
# Explorer
requests>=2.18.4

# CFG graphical view
graphviz>=0.9

# bytecode analytics
matplotlib>=2.1.1
numpy>=1.14.2

# ETH
# ethereum>=2.3.0
z3-solver==4.8.12

# WASM / EOS
wasm==1.2

# DWARF debug info
pyelftools==0.27

# Levenshtein
python-Levenshtein

# timeout decorator
timeout-decorator==0.5.0

# disjoint-set (https://github.com/mrapacz/disjoint-set)
disjoint_set==0.7.3

# cache tools
cachetools==5.0.0
